requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
multidict==6.0.4
pyaes==1.6.1
Pyrogram==2.0.70
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
TgCrypto==1.2.5
tzdata==2023.3
tzlocal==4.3
uvloop==0.17.0
yarl==1.9.2
